[#3830] Move chevah.emprical to chevah.compat.testing.
