+++ chevah/compat/tests/__init__.py
@@ -6,7 +6,7 @@ Tests for the system compatibility module.
 from __future__ import print_function
 from __future__ import division
 from __future__ import absolute_import
-from chevah.empirical import mk
+from chevah.compat.testing import mk
 
 
 def setup_package():
