+++ chevah/compat/tests/domain/__init__.py
@@ -7,7 +7,9 @@ Controller.
 from __future__ import print_function
 from __future__ import division
 from __future__ import absolute_import
-from chevah.empirical.testcase import (
+import os
+
+from chevah.compat.testing.testcase import (
     ChevahTestCase,
     )
 from chevah.compat.testing import (
@@ -29,9 +31,14 @@ def runDomainTest():
     """
     Return True if buildslave is a domain member.
     """
+    # FIXME:3832:
+    # Domain tests are broken.
+    return False
+
     # For now, elevated tests are executed only on the domain controller
     # buildslave.
-    if '-dc-' in ChevahTestCase.getHostname():
+    BUILDER_NAME = os.getenv('BUILDER_NAME', '')
+    if '-dc-' in BUILDER_NAME:
         return True
 
     return False
